fix: rename type alias, rename folder, css & data
