+++ practice-one/src/hooks/useFetch.tsx
@@ -48,7 +48,7 @@ function useFetch<T>(
     }
   }, [url, options, recallTrigger])
 
-  return { data, error, loading }
+  return { data, error, loading } as const
 }
 
 export default useFetch
